requirements.txt

httpretty
httplib2
lxml
setuptools
kostalpiko
#kostalpyko
paho-mqtt


